requirements.txt
absl-py==0.1.11
astor==0.6.2
bleach==1.5.0
captcha==0.2.4
certifi==2018.1.18
chardet==3.0.4
click==6.7
cycler==0.10.0
decorator==4.2.1
Flask==1.0.2
gast==0.2.0
grpcio==1.10.0
h5py==2.7.1
html5lib==0.9999999
idna==2.6
itsdangerous==0.24
Jinja2==2.10
Keras==2.1.6
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
networkx==2.1
numpy==1.14.2
pandas==0.22.0
Pillow==5.0.0
protobuf==3.5.2
pycryptodome==3.6.1
pymongo==3.6.1
pyparsing==2.2.0
pytesseract==0.2.0
python-dateutil==2.7.0
pytz==2018.3
PyWavelets==0.5.2
PyYAML==3.12
requests==2.18.4
schedule==0.5.0
scikit-image==0.13.1
scipy==1.0.1
six==1.11.0
tensorboard==1.6.0
tensorflow==1.6.0
termcolor==1.1.0
urllib3==1.22
Werkzeug==0.14.1
xmltodict==0.11.0
